fix: rename constants file, add private constructor
