+++ passengerService/src/main/java/com/modsen/passengerservice/dto/PassengerRequestDto.java
@@ -1,6 +1,6 @@
 package com.modsen.passengerservice.dto;
 
-import com.modsen.passengerservice.configuration.MessageConstants;
+import com.modsen.passengerservice.util.AppConstants;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -16,6 +16,6 @@ public record PassengerRequestDto (
     String email,
 
     @NotBlank(message = "{passenger.phone.mandatory}")
-    @Pattern(regexp = MessageConstants.PHONE_REGEXP, message = "{passenger.phone.invalid}")
+    @Pattern(regexp = AppConstants.PHONE_REGEXP, message = "{passenger.phone.invalid}")
     String phone) {
 }
